+++ helper_functions/internet_connection.py
