+++ helper_functions/scoreboard_helpers.py
@@ -2,13 +2,8 @@
 
 import FreeSimpleGUI as sg  # type: ignore import warning
 from settings import *
-from get_data.get_team_logos import get_random_logo
-import math
 import settings
 import time
-import platform
-from get_data.get_team_league import append_team_array
-from main import set_screen
 import subprocess
 import sys
 import gc
@@ -16,118 +11,9 @@ import tkinter as tk
 from tkinter import font as tkFont
 
 
-def gui_setup() -> sg.Window:
-    """Create General User Interface."""
-
-    sg.theme("black")
-    set_screen()  # Set the screen to display on
-    append_team_array(settings.teams)  # Get the team league and sport name
-    files = get_random_logo()
-
-    window_width = sg.Window.get_screen_size()[0]
-    window_height = sg.Window.get_screen_size()[1]
-
-    column_width = window_width / 3
-    column_height = window_height * .66
-    info_height = window_height * (1/6.75)
-    space_between_score = column_width / 8
-
-    print(f"\n\nWindow Width: {math.ceil(window_width)}, Window Height: {math.ceil(window_height)}")
-    print(f"Column Width: {math.ceil(column_width)}, Column Height: {math.ceil(column_height)}")
-    print(f"Info Height: {math.ceil(info_height)}")
-    print(f"Space Between Score: {math.ceil(space_between_score)}")
-
-    home_logo_layout = [
-        [sg.Push()],
-        [sg.VPush()],
-        [sg.Image(f"images/sport_logos/{files[0][0]}/{files[0][1]}.png", key='home_logo', pad=((0, 0), (0, 0)))],
-        [sg.VPush()],
-        [sg.Push()],
-    ]
-
-    away_logo_layout = [
-        [sg.Push()],
-        [sg.VPush()],
-        [sg.Image(f"images/sport_logos/{files[1][0]}/{files[1][1]}.png", key='away_logo', pad=((0, 0), (0, 0)))],
-        [sg.VPush()],
-        [sg.Push()],
-    ]
-
-    away_record_layout = [
-        [sg.Push()],
-        [sg.Text("AWAY", font=(settings.FONT, settings.RECORD_TXT_SIZE), key='away_record', pad=((0, 0), (0, 0)))],
-        [sg.Push()],
-    ]
-    home_record_layout = [
-        [sg.VPush()],
-        [sg.Push()],
-        [sg.Text("HOME", font=(settings.FONT, settings.RECORD_TXT_SIZE), key='home_record', pad=((0, 0), (0, 0)))],
-        [sg.Push()],
-    ]
-
-    above_score_layout = [
-        [sg.VPush()],
-        [sg.Push(), sg.Text("", font=(settings.FONT, settings.TOP_TXT_SIZE), key='above_score_txt', pad=((0, 0), (space_between_score, 0))), sg.Push()],
-        [sg.VPush()],
-    ]
-
-    score_layout = [
-        [sg.Text("SCO", font=(settings.FONT, settings.SCORE_TXT_SIZE), key='away_score', pad=((0, 0), (space_between_score, 0))),
-         sg.Text("-", font=(settings.FONT, settings.HYPHEN_SIZE), key='hyphen', pad=((0, 0), (space_between_score, 0))),
-         sg.Text("RE", font=(settings.FONT, settings.SCORE_TXT_SIZE), key='home_score', pad=((0, 0), (space_between_score , 0)))],
-        [sg.Text("", font=(settings.FONT, settings.TIMEOUT_SIZE), key='away_timeouts', pad=((0, 50), (0 , 25))),
-         sg.Text("", font=(settings.FONT, settings.TIMEOUT_SIZE), key='home_timeouts', pad=((50, 0), (0 , 25)))],
-    ]
-
-    below_score_image = [
-        [sg.VPush()],
-        [sg.Image("", key='under_score_image')],
-        [sg.VPush()],
-    ]
-
-    top_info_layout = [[sg.VPush()], [sg.Push(), sg.Text("", font=(settings.FONT, settings.NBA_TOP_INFO_SIZE), key='top_info'), sg.Push()]]
-    bottom_info_layout = [[sg.VPush()], [sg.Push(), sg.Text("Fetching Data...", font=(settings.FONT, settings.INFO_TXT_SIZE), key='bottom_info'), sg.Push()],[sg.VPush()],[sg.Push()]]
-
-    layout = [
-    [
-        sg.Column([  # Vertical stack for away team
-            [sg.Frame("", away_logo_layout, element_justification='center', border_width=0, size=(column_width, column_height * (4/5)))],
-            [sg.Frame("", away_record_layout, element_justification='center', border_width=0, size=(column_width, column_height * (1/5)))]
-        ], element_justification='center', pad=((0, 0), (0, 0))),
-        sg.Column([  # Vertical score
-            [sg.Frame("", above_score_layout, element_justification='center', border_width=0, size=(column_width, column_height * (1/4)))],
-            [sg.Frame("", score_layout, element_justification='center', border_width=0, size=(column_width, column_height * (7/16)))],
-            [sg.Frame("", below_score_image, element_justification='center', border_width=0, size=(column_width, column_height * (5/16)))]
-        ], element_justification='center', pad=((0, 0), (0, 0))),
-        sg.Column([  # Vertical stack for home team
-            [sg.Frame("", home_logo_layout, element_justification='center', border_width=0, size=(column_width, column_height * (4/5)))],
-            [sg.Frame("", home_record_layout, element_justification='center', border_width=0, size=(column_width, column_height * (1/5)))]
-        ], element_justification='center', pad=((0, 0), (0, 0))),
-    ],
-        [sg.VPush()],
-        [sg.Frame("", top_info_layout, element_justification='center', border_width=0, size=(window_width, info_height * (6/7)))],
-        [sg.Frame("", bottom_info_layout, element_justification='center', border_width=0, size=(window_width, info_height))],
-        [sg.Frame("",
-                  [[sg.Push(), sg.Text("Created by: Matthew Ferretti", font=(settings.FONT, settings.SIGNATURE_SIZE), key="signature")]],
-                  element_justification='bottom', border_width=0, size=(window_width, info_height * (1/7)))],
-    ]
-
-    # Create the window
-    window = sg.Window("Scoreboard", layout, no_titlebar=False, resizable=True, return_keyboard_events=True).Finalize()
-    
-    # Maximize does not work on MacOS, so we use attributes to set fullscreen
-    if platform.system() == 'Darwin':
-        window.TKroot.attributes('-fullscreen', True)
-    else:
-        window.Maximize()
-    window.TKroot.config(cursor="none")  # Remove cursor from screen
-
-    return window
-
-
 def will_text_fit_on_screen(text: str) -> bool:
     """Check if text will fit on screen.
-    
+
     :param text: str to compare to width of screen
 
     :return bool: boolean value representing if string will fit on screen
@@ -146,9 +32,10 @@ def will_text_fit_on_screen(text: str) -> bool:
     else:
         return False
 
+
 def reset_window_elements(window: sg.Window) -> None:
     """Reset window elements to default values.
-    
+
     :param window: element that can be updated for displaying information
     """
     window['top_info'].update(value='', font=(FONT, settings.NOT_PLAYING_TOP_INFO_SIZE), text_color='white')
@@ -165,7 +52,7 @@ def reset_window_elements(window: sg.Window) -> None:
 
 def check_events(window: sg.Window, events, currently_playing=False) -> None:
     """Check for specific key presses.
-    
+
     :param window: element that can be updated for displaying information
     :param events: key presses that were recorded
     :param currently_playing: current state of scoreboard allowing for more or less key presses
@@ -229,7 +116,7 @@ def set_spoiler_mode(window: sg.Window, team_info: dict) -> sg.Window:
     window["top_info"].update(value="Will Not Display Game Info", font=(settings.FONT, settings.MLB_BOTTOM_INFO_SIZE))
     window['bottom_info'].update(value="No Spoiler Mode On", font=(settings.FONT, settings.MLB_BOTTOM_INFO_SIZE))
     window["under_score_image"].update(filename='')
-    if "@" not in team_info["above_score_txt"]: # Only remove if text doesn't contain team names
+    if "@" not in team_info["above_score_txt"]:  # Only remove if text doesn't contain team names
         window["above_score_txt"].update(value='')
     window["home_score"].update(value='0', text_color='white')
     window["away_score"].update(value='0', text_color='white')
