+++ screens/clock_screen.py
@@ -7,8 +7,8 @@ import time
 import FreeSimpleGUI as sg  # type: ignore import warning
 from get_data.get_team_logos import get_random_logo
 from get_data.get_espn_data import get_data
-from gui_setup import reset_window_elements
-from internet_connection import is_connected, reconnect
+from helper_functions.scoreboard_helpers import reset_window_elements
+from helper_functions.internet_connection import is_connected, reconnect
 from adafruit_ticks import ticks_ms, ticks_add, ticks_diff  # type: ignore import warning
 import settings
 import gc
